Filter out RQA messages labelled as Noise_Other_Project
