+++ pipeline.py
@@ -173,8 +173,8 @@ if __name__ == "__main__":
      print("Exporting advert CSV...")
      advert_phone_numbers = AdvertPhoneNumbers.generate(data, phone_number_uuid_table, advert_phone_numbers_csv_output_path)
 
-     print("Filtering out RQA and Follow up Messages labelled as Noise_Other_Project...")
-     data = FilterNOP.filter_noise_other_project(data)
+     print("Filtering out RQA Messages labelled as Noise_Other_Project...")
+     data = FilterNOP.filter_rqa_noise_other_project(data)
 
      print("Generating Analysis CSVs...")
      data = AnalysisFile.generate(user, data, csv_by_message_output_path, csv_by_individual_output_path)
