+++ src/__init__.py
@@ -6,4 +6,4 @@ from .production_file import ProductionFile
 from .apply_manual_codes import ApplyManualCodes
 from .analysis_file import AnalysisFile
 from .advert_phone_numbers import AdvertPhoneNumbers
-from .filter_noise_other_project import FilterNOP
+from .filter_rqa_noise_other_project import FilterNOP
